// src/app/modules/Faculty/faculty.interface.ts
import { Model, Types } from 'mongoose'

export type TGender = 'male' | 'female' | 'other'
export type TuserName = {
  firstName: string
  middleName: string
  lastName: string
}
export type TBloodGroup =
  | 'A+'
  | 'A-'
  | 'B+'
  | 'B-'
  | 'AB+'
  | 'AB-'
  | 'O+'
  | 'O-'

export type TFaculty = {
  id: string
  user: Types.ObjectId
  designation: string
  name: TuserName
  gender: TGender
  dateOfBirth?: Date
  email: string
  contactNo: string
  emergencyContactNo: string
  bloodGroup?: TBloodGroup
  presentAddress: string
  permanentAddress: string
  profileImg?: string
  academicDepartment: Types.ObjectId
  academicFaculty: Types.ObjectId
  isDeleted: boolean
}

export interface FacultyModel extends Model<TFaculty> {
  isUserExists(id: string): Promise<TFaculty | null>
}